aca_sft/requirements.txt
azure-storage-blob>=12.23.0,<13
datasets==4.4.2
pyarrow==22.0.0
python-dotenv>=1.0.1,<2
transformers==4.57.3
